backend/requirements.txt
Django>=4.2,<6
Pillow>=10,<11
gunicorn
dj-database-url
psycopg2-binary
whitenoise
python-dotenv
celery
redis
django-celery-beat
django-environ
cloudinary
django-cloudinary-storage
django-anymail
drf-spectacular
django-cors-headers
djangorestframework
djangorestframework-simplejwt
dj-rest-auth
requests
